Remove dpkg hold selection before installing Netdata
It seems that DPKG got more restrictive or the Ansible apt module was bugged before. Anyway, without this task, the upgrade of Netdata on recent Debian versions failed because the "hold" marking of DPKG was stronger than forcing the update.

With this commit, we first unhold the package before upgrading it.