Add Netdata repository with template
If you previously used the stable version of Netdata and then switched to nightlies, with `apt_repository`, you have both the regular and edge repository in your list. Those two repositories in the list cause troubles when switching back to stable, as edge always has newer versions available for the plugin packages (like the go collector plugin).

Using the template solves this issue, as only one source will be present.