Specify plugin list for Netdata
Netdata announced that they split their package into multiple small ones:
https://blog.netdata.cloud/split-plugin-packages/

This change resulted in some troubles for us. As we like to specify an exact version number for Netdata, but apt always tried to get the latest version, it usually resulted in an error like this:

```
The following packages have unmet dependencies:
 netdata-plugin-apps : Depends: netdata (= 1.40.0-48-nightly) but 1.40.0-45-nightly is to be installed
                       Conflicts: netdata (< 1.40.0-48-nightly) but 1.40.0-45-nightly is to be installed
```

By specifying and locking the version of all packages, the dependency resolution works correctly.

Sidenote: It appears that all packages are required for now, contractiding the blog post:

```
$ sudo apt-cache depends netdata
netdata
  ...
  Depends: netdata-plugin-ebpf
  Depends: netdata-plugin-apps
  Depends: netdata-plugin-pythond
  Depends: netdata-plugin-go
  Depends: netdata-plugin-debugfs
  Depends: netdata-plugin-nfacct
  Depends: netdata-plugin-chartsd
  Depends: netdata-plugin-slabinfo
  Depends: netdata-plugin-perf
```